maint: move complex-datatypes.store into dedicated folder
